requirements.txt
flask-sqlalchemy
psycopg2
flask-wtf
flask-ckeditor
python-dotenv

